+++ FrontEnd/src/pages/ProfileDetail/DetailedInfo/DetailedInfoComponents/Company.module.css
@@ -2,7 +2,6 @@
     display: flex;
     flex-direction: column;
     align-items: flex-start;
-    max-width: 343px;
 }
 
 .about-company__content--advantage,
@@ -10,7 +9,6 @@
     display: flex;
     flex-direction: column;
     align-items: flex-start;
-    gap: 12px;
 }
 
 .about-company__content--advantage--text,
