+++ FrontEnd/src/pages/ProfileDetail/DetailedInfo/DetailedInfoComponents/CompanyDescription.module.css
@@ -2,6 +2,4 @@
     display: flex;
     flex-direction: column;
     align-items: flex-start;
-    gap: 16px;
-    min-width: 800px;
 }
