+++ FrontEnd/src/pages/ProfileDetail/DetailedInfo/DetailedInfoComponents/Logistics.module.css
@@ -1,29 +1,33 @@
-.products-services {
+.logistics-wrapper {
     display: flex;
     flex-direction: column;
     align-items: flex-start;
-    gap: 12px;
     border-radius: 8px;
 }
 
-.products-services__title {
+.logistics,
+.logistics__content--block {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+}
+
+.logistics__title {
     display: flex;
     flex-direction: column;
     align-items: flex-start;
     align-self: stretch;
 }
 
-.products-services__title--block {
+.logistics__title--block {
     display: flex;
     padding: 8px 12px;
     align-items: flex-start;
-    gap: 10px;
     align-self: stretch;
     border-radius: 4px;
 }
 
-.products-services__title--text {
-    width: 742px;
+.logistics__title--text {
     color: var(--main-grey-90, #25292C);
     font-family: var(--font-main);
     font-size: 16px;
@@ -33,40 +37,10 @@
     text-transform: uppercase;
 }
 
-.products-services__title--divider {
-    height: 1px;
-    width: 795px;
-    background: var(--main-grey-20, #DEE1E8);
-}
-
-.products-services__content {
+.logistics__content--description {
     display: flex;
     padding: 12px;
-    flex-direction: column;
     align-items: flex-start;
-    gap: 24px;
-}
-
-.products-services__content--block {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 12px;
-}
-
-.products-services__content--title {
-    color: var(--main-grey-90, #25292C);
-    font-feature-settings: 'calt' off;
-    font-family: var(--font-main);
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 600;
-    line-height: 20px;
-    letter-spacing: -0.16px;
-}
-
-.products-services__content--description {
-    width: 776px;
     color: var(--main-grey-90, #25292C);
     font-feature-settings: 'calt' off;
     font-family: var(--font-main);
