+++ FrontEnd/src/pages/ProfileDetail/DetailedInfo/DetailedInfoComponents/ProductsServices.module.css
@@ -1,37 +1,26 @@
-.cooperation-wrapper {
+.products-services {
     display: flex;
     flex-direction: column;
     align-items: flex-start;
-    gap: 24px;
     border-radius: 8px;
 }
 
-.cooperation,
-.cooperation__content--block {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 12px;
-}
-
-.cooperation__title {
+.products-services__title {
     display: flex;
     flex-direction: column;
     align-items: flex-start;
     align-self: stretch;
 }
 
-.cooperation__title--block {
+.products-services__title--block {
     display: flex;
     padding: 8px 12px;
     align-items: flex-start;
-    gap: 10px;
     align-self: stretch;
     border-radius: 4px;
 }
 
-.cooperation__title--text {
-    width: 742px;
+.products-services__title--text {
     color: var(--main-grey-90, #25292C);
     font-family: var(--font-main);
     font-size: 16px;
@@ -41,18 +30,31 @@
     text-transform: uppercase;
 }
 
-.cooperation__title--divider {
-    height: 1px;
-    width: 795px;
-    background: var(--main-grey-20, #DEE1E8);
-}
-
-.cooperation__content--description {
+.products-services__content {
     display: flex;
     padding: 12px;
+    flex-direction: column;
     align-items: flex-start;
-    gap: 10px;
-    width: 776px;
+}
+
+.products-services__content--block {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+}
+
+.products-services__content--title {
+    color: var(--main-grey-90, #25292C);
+    font-feature-settings: 'calt' off;
+    font-family: var(--font-main);
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: 20px;
+    letter-spacing: -0.16px;
+}
+
+.products-services__content--description {
     color: var(--main-grey-90, #25292C);
     font-feature-settings: 'calt' off;
     font-family: var(--font-main);
